fix bug and rename reducer
