+++ packages/client/src/constants/constantValues.js
@@ -1,8 +1,8 @@
 const DEFAULT_INCLUSION_EFFORT = "1.00000000";
 const EXECUTION_EFFORTS = {
-  PROPOSE_ACTION: 0.00000074,
-  SIGN_ACTION: 0.00000061,
-  EXECUTE_ACTION: 0.00000084,
+  PROPOSE_ACTION: "0.00000074",
+  SIGN_ACTION: "0.00000061",
+  EXECUTE_ACTION: "0.00000084",
 };
 
 export { DEFAULT_INCLUSION_EFFORT, EXECUTION_EFFORTS };
