--- packages/client/src/reducers/index.js
@@ -1,4 +1,0 @@
-export { default as treasuries } from "./treasuries";
-export { default as nfts } from "./nfts";
-export { default as vaults } from "./vaults";
-export { default as contacts } from "./contacts";
